Controversial: simplify Digit to allow for ergonomics
Problem: right now, to create a Digit manually, we need to always
pass in a doesParse proof of e.g. form 'Exists.intro {} $ by trivial',
as it's difficult to force as a default argument. However, why is
'doesParse' prop even of this form
'∃ arg : Cached parseDigit x, Option.isSome arg.val'? We already store
'parsed', itself of the form 'Cached parseDigit x', in the same struct.
Considering Cached is supposed to be a singleton, _if_ the arg exists,
it's going to be equal to 'parsed'. Thus, we only need to check that
'parsed' itself isSome! This will also allow us to set a default proof.

Solution: removed the existential quantifier from Digit's doesParse.