Turn the Const getter into an op
Problem: currently, e.g. i32.const is parsed as a getter, but it's
actually an instruction/operation which pushes a value onto the stack.
It doesn't make much sense as a getter either, as getting its value
for e.g. i32.add is covered by both the from_stack and from_operation.

Solution: changed it into an op. This has the added benefit that we can
now parse/process functions whose bodies are just e.g. '(i32.const 0)'.