[#30] Support labels on stack
Problem: to support proper branching with e.g. `br_if`, we need to
introduce labels to our `block` and `loop` structured instructions
to allow branch instructions to target them.

Solution: added another type of stack entries, namely labels, as well
as defined a LabelIndex structure for further work. Came up with how
to pass labels to the inner operations even when the whole value stack
shouldn't be available to them. This opens up the path to actually
implement `br` and `br_if`.