Add StateT to carry stack around
Problem: in the engine, we currently pass in the stack as an explicit
argument to every worker function invocation. It works fine, but is prone
to mistakes when modifying the stack sequentially, like in binops, and
hurts readability.

Solution: added 'StateT' monad transformer to the engine type. It carries
the 'List StackEntry' that we're using as a stack. Defined an abbreviation
'EngineM' that should house all monads we might want to stack further.