Add ExceptT for easy continuation passing
Problem: continuation handling is currently a real pain. Basically,
we need to check that we don't have a continuation after every single
'runOp' ever run, because otherwise we risk missing a branch/jump.
This results in a code that is massively ridden with boilerplate and
hardly readable too.

Hold on, but in the control flow sense, continuations are almost like
exceptions!

Solution: added 'ExceptT (List Operation)' to our 'EngineM' monad stack
abbreviation. 'ExceptT' on top allows us to still access the stack state
and, consequently, to recover (=execute the continuation) nicely.
Also followed through with refactoring that this change allows.