Add preliminary support for some float operations
Namely eq, ne, gt, lt, ge, le, add, sub, mul.

Problem: we can't use Lean floats in Wasm arithmetics because of
opaqueness and of instances not being defined with bit representation
in mind. The main problem even for `eq` is that only BEq is defined,
according to which +0==-0, which is a behaviour Wasm contradicts.
Also, consequently there isn't a way to implement `copysign`.

Currently there is no solution, but down the road we should either
get rid of floats support in Wasm.lean altogether (bad for the
completeness goal) or implement our own Floats.