Get rid of Get
Problem: at the very start of Wasm.lean implementation, we defined two
getters: Get and Get'. Get made use of internal Wasm types for dependent
typing, Get' was a wrapper on top of it to actually allow for use in the
engine and for getters on top of operations.

However, with the evolution of Wasm.lean two problems have arisen:
1. we have sadly moved away from dependent typing for Wasm.lean, and this
means we don't have an easy way to come up with types for Get to check upon
parsing.
2. the initial design had consts and locals as inherent getters. However,
both are actually instructions, and it's possible to use them on their own.
Const has already been cut out and turned into an instruction, and now it's
time to do that for locals for proper support - but that means the underlying
Get is reduced to just the .from_stack constructor. Get doesn't really type
check stack entries, so the idea is lost anyway.

Solution: removed the Get inductive and flattened the parser structure to
operate on pure Get'. Get' is now ready to be reduced to just .from_stack
and .from_operation.