Bring locals into the monad stack
Problem: we still pass locals around as an explicit parameter, which won't
nicely work any longer now that we need to modify the locals list. We
would have to return the possibly modified list with each runOp.

Solution: introduced another StateT layer to our EngineM stack. The nice
thing about Lean is that, unlike Haskell, it supports multiple StateTs in
a stack no problem because there's no functional dependency for the state
type by default.
Note: the locals aren't being modified or indeed accessed yet, that's in
the next commit.

The new StateT layer houses a List of local values, which have changed a bit.
Whereas before we have used (Option Name x Option StackEntry) to simulate
a store of Locals, this isn't actually sufficient for type checking:
additionally, we need a type specified for this param/local. Rather than
introduce a third field for this product, which wouldn't be very readable,
we define a new LocalEntry structure.