Bytes.lean: hide locals and globals under ReaderT
Problem: sometimes when extracting, but only when dealing with local
or global variable instructions, we need to have an indexed list of
locals/globals to look up a name of a global. This forces us to face
the same problem as in the Engine.lean: passing around these lists
explicitly is quickly becoming a hassle.

However, unlike in the Engine, we never need to modify these lists.

Solution: sounds like a ReaderT! Defined ExtractM, an abbrev for two
ReaderTs, housing Globals and Locals needed for extraction. Also defined
some auxiliary HAppend instances to leave most of the code unchanged.