Remove indices from Local and Global definition
Problem: right now, both the Local and Global structure have a field
'index : Nat', mean to represent the index that variable instructions,
e.g. 'local.get', use to access it. With this approach, it needs to be
calculated basically at parse time. This is not only unwieldy in itself,
it's also semantically problematic. For example, for globals, indices
only start from the first global definition that isn't an import.
Moreover, we only use indices in the Engine (and binary encoding), so
semantically it would make sense to index them there.

Solution: removed indices from the two definitions. They weren't used
much anywhere else as is.