[#46] Support ids for structured control instructions
Problem: per the spec, structured control instructions, i.e.
`block`, `loop`, and `if`, support binding optional symbolic
identifiers, similar to e.g. functions. We don't support this yet,
but it would be useful and good for UX.

Solution: added an optional string name field to block, loop, if
operations. Supported them in the parser and in the engine. Supported
overshadowing labels (or rather, according to the spec, overwriting).
Added a StateM monad to ExtractM in Bytes.lean to support proper
encoding of branch instructions using symbolic identifiers.
Added parse and binary encoding tests to ensure correctness.

Closes #46.