Revamp whitespace parsing
Problem: we don't fully adhere to the spec when parsing whitespace.
Currently, we only ignore spaces, tabs, and newlines, but parse \r
as a valid character and can't parse comments at all.

Solution: defined line and block comment parsers, changed the
whitespace parsers to be consistent with the spec. Added some tests.