Bring running functions into EngineM
Problem: we have no way to run functions from inside the EngineM,
meaning we can't recursively invoke/'call' functions from inside
another function. The only run method we have, 'runDo', is meant
for "outside" usage.

Solution: reused the 'runDo' code in a new function for EngineM,
'runFunc' and cleaned up the predecessors.
Note that currently it's still single-function only, but it's a start.