Carry the whole Store around in EngineM
Problem: to implement 'call'/other function invocations, we need
a way to access other function instances. Per the specification,
this is done by:
- storing an activation frame on stack that points to the function
  instance;
- having the function instance depend/carry around a module instance;
- having the module instance, in turn, point to all function instances.

The astute reader already realises that for Lean, this is kinda problematic.
There is a circular dependency between FunctionInstance and ModuleInstance
that isn't easily solvable: Lean only supports mutual declarations for
inductives, not for structures, and we would really really like to keep
structures instead of one-constructor inductives for accessibility.

Solution: per spec, module instances are actually indices into the store.
So why not just carry the store around, as we would need to anyway, and
leave function instances separate and module instances non-existing?

Brought 'Store m' into the EngineM monad stack, replacing Global instances,
which have to reside in the store anyhow.