Bring activation frames into EngineM
Problem: per the specification, local variables are stored on-stack,
in an 'activation frame' stack entry, one-per-function invocation.
This is, of course, unwieldy and very inefficient for us (as we
currently use a List to model the stack), so instead, we carry locals
around in the monadic stack. Now, as we get closer to 'call'
implementation, we also need the function arity that is also stored
in the same activation frame construct.

Solution: changed the StateT Locals to StateT Frame, ensuring the
locals' usage doesn't change in practice.