Get rid of Get'
Problem: once upon a time, we have chosen an architecture/design where
we have a predetermined amount of operands for each defined operation,
e.g. two operands for `i32.add`, gotten either from the stack or from
another "nested" operation. This turns out to be not exactly the greatest
design choice, as WA(S)T, in all its glory, allows instructions to be
nested completely arbitrarily, including in operations that don't touch
the stack themselves. Actually, this is specced as "folded instructions",
which is a concept we currently can't support at all.

Solution: removed getters entirely. Now we just follow 6.5.10 of the spec,
which postulates that these folded instructions are just unfolded under
the hood, putting the "main" intruction after all the "nested" instructions.

Note this makes us lose the original representation at the parsing stage,
i.e. we wouldn't be able to round-trip test the pretty-printing, but I'm
willing to live with it at this stage.

Also added some tests to RuntimeCompatibility that were used in filing
previous issues concerned with this behaviour. Just so I can safely say:

Closes #62, closes #63